Draft of my part
First of all, I defined a set as an item containing at least 2 parts. I cleaned the data by filtering out rows with fewer parts than 2.  Then, I explored the distribution of number of parts per set in a histogram and number of parts per set throughout the years in a scatter plot. Lastly, I presented the average number of parts in a set.